+++ app/src/main/java/io/bloco/circles/ui/auth/identities/IdentitiesViewModel.kt
@@ -8,6 +8,7 @@ import io.bloco.circles.data.IdentityModel
 import io.bloco.circles.data.interfaces.IdentityRepository
 import io.bloco.circles.domain.GenerateAuthResponse
 import io.bloco.circles.domain.GetAppDetails
+import io.bloco.circles.domain.NewIdentity
 import io.bloco.circles.ui.BaseViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
@@ -24,13 +25,16 @@ class IdentitiesViewModel
     private val authRequestsStore: AuthRequestsStore,
     generateAuthResponse: GenerateAuthResponse,
     identityRepository: IdentityRepository,
-    getAppDetails: GetAppDetails
+    getAppDetails: GetAppDetails,
+    newIdentity: NewIdentity,
 ) : BaseViewModel() {
 
     // Inputs
     private val identitySelected = BroadcastChannel<IdentityModel>(1)
+    private val createNewIdentity = BroadcastChannel<Unit>(1)
 
     // Outputs
+    private val chooseNewIdentityUsername = BroadcastChannel<Unit>(1)
     private val sendAuthResponse = BroadcastChannel<AuthResponseModel>(1)
     private val identities = BroadcastChannel<List<IdentityModel>>(1)
     private val appDetails = MutableStateFlow<AppDetails?>( null)
@@ -41,7 +45,6 @@ class IdentitiesViewModel
     init {
         identityRepository
             .observe()
-            .take(1)
             .onEach {
                 identities.send(it)
             }
@@ -56,6 +59,21 @@ class IdentitiesViewModel
                 }
             }
         }
+
+        createNewIdentity
+            .asFlow()
+            .onEach {
+                authRequestsStore.get()?.let { request ->
+                    if(request.registerSubdomain) {
+                        chooseNewIdentityUsername.send(Unit)
+                    } else {
+                        loading.send(true)
+                        newIdentity.create()
+                        loading.send(false)
+                    }
+                }
+            }
+            .launchIn(ioScope)
 
         identitySelected
             .asFlow()
@@ -82,8 +100,10 @@ class IdentitiesViewModel
 
     // Inputs
     suspend fun identitySelected(identity: IdentityModel) = identitySelected.send(identity)
+    suspend fun createNewIdentity() = createNewIdentity.send(Unit)
 
     // Outputs
+    fun chooseNewIdentityUsername(): Flow<Unit> = chooseNewIdentityUsername.asFlow()
     fun sendAuthResponse(): Flow<AuthResponseModel> = sendAuthResponse.asFlow()
     fun identities(): Flow<List<IdentityModel>> = identities.asFlow()
     fun appDetails(): Flow<AppDetails?> = appDetails.asStateFlow()